Update stylus, reset styles, and various pug templates
This commit includes a substantial enhancement to the "front.styl" file, tweaks to the "reset.styl", and various adjustments to many .pug templates. Also, "jsScripts" importation in "gulpfile.js" has been uncommented and this resulted in some changes in the exported tasks. The usage of SVG icons were also modified, particularly